Attempt to refactor various draw.c APIs into a new vfunc table "dapi"
This is so individual functions can potentially replaced for certain
renderers / exporters, without requiring to re-implement all of the
draw.c rendering code each time.

For example - this might allow the GL renderer to re-implement the
layer drawing function so as to render "mask first" into the stencil
buffer.

XXX: Not properly updated for BB vias yet