Fix rotation of components on back-side of board
The rotate action always rotates in a given direction (CCW or CW), viewed
from the top of the PCB. Since we keep the rotation attribute fixed when
flipping the component to the back side of the board, this implies that the
rotation angle is viewed from whichever side of the board the component
resides on, and the delta we apply to the "PCB::rotation" attribute must
depend on which side the component is on.

An alternative possible approach, where rotation is always described
relative to the top side of the board would require alteration of the
transforms applied to 3D models, but also a correction to rotation angle
when parts are flipped from one side to the other. Fixing it in this way
would break existing boards (assuming parts rotated on the top-side have
been flipped to the bottom of the PCB).