Fix timeout forced redraw to run from the main loop, not directly
This avoids us drawing when the board is in an intermediate state. Will
likely not give as much immediacy as drawing from where the graphics are
invalidated - but avoids visual artifacts when drawing tacks with
auto-enforce DRC enabled.

This should be credited to Peter Clifton, not me. I screwed up the
index.