WIP: Move DRC code to drc.[c,h]
A first step towards modularizing the DRC code. It doesn't build yet,
hence the WIP. There are some issues sorting out some of the global
variables, like "drc", and "User". In some places the User variable is
set and reset by the code remaining in find.c, but this variable is
supposed to indicate that the DRC check was initiated by the user. So,
that doesn't really make sense. Some time needs to be spent documenting
this and trying to figure out why things are executed the way they are.

Peter C had done a little bit of work in this direction, and some of
that has been cherry-picked. There's some additional stuff to consider,
the "store_undo" stuff. This might help the "User" variable to go away,
since the undoing is presumably to restore the state of flags since the
DRC check wipes out those states.