DRC: Multiple enhancements
This commit does several things. All violations of a kind connected to a
pin or via are now flagged, instead of stopping after the first. Also,
both violating objects are now highlighted instead of just the source
object.

Some notes about this: I've changed how the DRCFLAG is used, and the
behavior of the find.c routines when the global "drc" variable is set.
(This variable should be renamed, probably) The DRCFLAG should be set on
the first pass of the connection building. On the second pass, the
DRCFLAG serves to propagate the search. Only objects for which the
DRCFLAG is set will be searched for overlaps. This allow all of the
violations to be found, without the search spidering out through the
entire design.

The "drc" global variable should be set on the second pass through the
connection building algorithm. When this is set, the search will return
early when it finds an object that hasn't been flagged already. It uses
the SELECTEDFLAG for this, although now it can probably be changed to
use the DRCFLAG instead. If a search is ended early, it can be continued
by calling DoIt again.